# openapi/models/response_result.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"bytes"
	"context"
	"encoding/json"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// ResponseResult response result
//
// swagger:model ResponseResult
type ResponseResult struct {

	// reliability
	Reliability int64 `json:"reliability,omitempty"`
}

// UnmarshalJSON unmarshals this object while disallowing additional properties from JSON
func (m *ResponseResult) UnmarshalJSON(data []byte) error {
	var props struct {

		// reliability
		Reliability int64 `json:"reliability,omitempty"`
	}

	dec := json.NewDecoder(bytes.NewReader(data))
	dec.DisallowUnknownFields()
	if err := dec.Decode(&props); err != nil {
		return err
	}

	m.Reliability = props.Reliability
	return nil
}

// Validate validates this response result
func (m *ResponseResult) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this response result based on context it is used
func (m *ResponseResult) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *ResponseResult) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *ResponseResult) UnmarshalBinary(b []byte) error {
	var res ResponseResult
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}