# openapi/restapi/operations/server/get_session_responses.go
// Code generated by go-swagger; DO NOT EDIT.

package server

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"net/http"

	"github.com/go-openapi/runtime"
	"github.com/go-openapi/runtime/middleware"

	"battle_of_psychics/openapi/models"
)

// GetSessionOKCode is the HTTP code returned for type GetSessionOK
const GetSessionOKCode int = 200

/*
GetSessionOK Success create session

swagger:response getSessionOK
*/
type GetSessionOK struct {

	/*
	  In: Body
	*/
	Payload *models.ResponseSession `json:"body,omitempty"`
}

// NewGetSessionOK creates GetSessionOK with default headers values
func NewGetSessionOK() *GetSessionOK {

	return &GetSessionOK{}
}

// WithPayload adds the payload to the get session o k response
func (o *GetSessionOK) WithPayload(payload *models.ResponseSession) *GetSessionOK {
	o.Payload = payload
	return o
}

// SetPayload sets the payload to the get session o k response
func (o *GetSessionOK) SetPayload(payload *models.ResponseSession) {
	o.Payload = payload
}

// WriteResponse to the client
func (o *GetSessionOK) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.WriteHeader(200)
	if o.Payload != nil {
		payload := o.Payload
		if err := producer.Produce(rw, payload); err != nil {
			panic(err) // let the recovery middleware deal with this
		}
	}
}

func (o *GetSessionOK) GetSessionResponder() {}

// GetSessionBadRequestCode is the HTTP code returned for type GetSessionBadRequest
const GetSessionBadRequestCode int = 400

/*
GetSessionBadRequest error create session

swagger:response getSessionBadRequest
*/
type GetSessionBadRequest struct {
}

// NewGetSessionBadRequest creates GetSessionBadRequest with default headers values
func NewGetSessionBadRequest() *GetSessionBadRequest {

	return &GetSessionBadRequest{}
}

// WriteResponse to the client
func (o *GetSessionBadRequest) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(400)
}

func (o *GetSessionBadRequest) GetSessionResponder() {}

// GetSessionInternalServerErrorCode is the HTTP code returned for type GetSessionInternalServerError
const GetSessionInternalServerErrorCode int = 500

/*
GetSessionInternalServerError internal error

swagger:response getSessionInternalServerError
*/
type GetSessionInternalServerError struct {
}

// NewGetSessionInternalServerError creates GetSessionInternalServerError with default headers values
func NewGetSessionInternalServerError() *GetSessionInternalServerError {

	return &GetSessionInternalServerError{}
}

// WriteResponse to the client
func (o *GetSessionInternalServerError) WriteResponse(rw http.ResponseWriter, producer runtime.Producer) {

	rw.Header().Del(runtime.HeaderContentType) //Remove Content-Type on empty responses

	rw.WriteHeader(500)
}

func (o *GetSessionInternalServerError) GetSessionResponder() {}

type GetSessionNotImplementedResponder struct {
	middleware.Responder
}

func (*GetSessionNotImplementedResponder) GetSessionResponder() {}

func GetSessionNotImplemented() GetSessionResponder {
	return &GetSessionNotImplementedResponder{
		middleware.NotImplemented(
			"operation authentication.GetSession has not yet been implemented",
		),
	}
}

type GetSessionResponder interface {
	middleware.Responder
	GetSessionResponder()
}