# openapi/swagger.yaml
swagger: "2.0"
info:
  description: "Service The battle of psychics"
  title: battle
  version: 1.0.0
consumes:
  - application/json
produces:
  - application/json
schemes:
  - http
securityDefinitions:
   key:
    type: apiKey
    in: header
    name: x-token
security:
  - key: []

paths:            
  /healthCheck:
    get:
      tags:
        - Standard
      security:
        - {}
      operationId: healthCheck
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              ok:
                type: boolean

  /Session:
    get:
      tags:
       - server
      security:
       - {}
      operationId: getSession
      responses:
        200:
          description: Success create session
          schema:
            $ref: "#/definitions/ResponseSession"
        400:
          description: error create session
        500:
          description: internal error

  /Conceive:
    get:
      tags:
        - server
      security:
        - key: []
      operationId: conceive
      responses:
        200:
          description: OK
          schema:
            type: array
            items: 
              $ref: "#/definitions/ResponseСonceive"
        401:
          description: incorrect api key auth
        404:
          description: a session with such a token was not found
        500:
          description: internal error

  /Result:
    post:
      tags:
       - server
      security:
        - key: []
      operationId: result
      parameters:
        - name: body
          in: body
          schema:
            $ref: "#/definitions/Result"
      responses:
        200:
          description: ok
          schema: 
            type: array
            items: 
              $ref: "#/definitions/ResponseResult"
        401:
          description: incorrect api key auth
        404:
          description: a session with such a token was not found
        406:
          description: cannot call a method without first calling /Conceive
        500:
          description: internal error

definitions:
  error:
    type: object
    required:
      - code
      - message
    properties:
      code:
        description: Either same as HTTP Status Code OR >= 600.
        type: integer
        format: int32
      message:
        type: string
  
  ResponseResult:
    type: object
    properties:
      reliability:
        type: integer
  
  ResponseСonceive:
    type: object
    properties:
      hypothesis:
        type: integer

  ResponseSession:
    type: object
    properties:
      jwt:
        type: string

  Result:
    type: object
    properties:
      number:
        type: integer

  principal:
    type: string